Refactoringclass
