+++ New Unity Project/Assets/Scenes/SampleScene.unity
@@ -2231,7 +2231,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 58.69201, y: -6.1143947}
+  m_AnchoredPosition: {x: 58.692017, y: -6.11438}
   m_SizeDelta: {x: 117.38402, y: 12.228789}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &97516938
@@ -9918,7 +9918,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 1807429813}
   m_Direction: 0
   m_Value: 0
-  m_Size: 1
+  m_Size: 0.99994755
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -18106,7 +18106,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 1}
   m_AnchorMax: {x: 0.5, y: 1}
-  m_AnchoredPosition: {x: -71.52498, y: 0.0000013877378}
+  m_AnchoredPosition: {x: -71.52499, y: 0.0000013877378}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &896579102
