+++ New Unity Project/Assets/Scripts/Querys.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36aa98d1b8c599f48b4664a864925192
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
