+++ New Unity Project/Assets/Scripts/Querys/QryElementos.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07270be59666d4241a0b4f95ef743f1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
