+++ New Unity Project/Assets/Scripts/Querys/QryMoleculas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 159479a6bef01084096c3a40b43b235a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
